feat: migrate to grpc (#62)
- new docker image, which does protobuf -> go code generation
- remove common/dapr.go in favor of common/grpc.go
- dapr sdk not used for now
- Logging, Auth and Validate Interceptors (Middleware) are available
- new Logging interface which allows a custom LoggableFields function for redaction purposes (respected in Logging middeware)
- services now use zerolog.Ctx(ctx) with the request's ctx
- minor changes here and there for compat reasons

+++ services/emergency-room-svc/models/er.go
@@ -10,7 +10,7 @@ type EmergencyRoomBase struct {
 	Location           hwgorm.Point `json:"location" gorm:"default:NULL" validate:"required"`
 	DisplayableAddress string       `json:"displayableAddress" gorm:"default:NULL" validate:"required"`
 	Open               bool         `json:"open" gorm:"column:is_open;default:true"`
-	Utilization        int16        `json:"utilization" gorm:"default:1"`
+	Utilization        int32        `json:"utilization" gorm:"default:1"`
 }
 
 type EmergencyRoom struct {
